fix(tools)!: removes the esbuild plugins and helpers
